% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarization.R
\name{csv_summary}
\alias{csv_summary}
\title{Apply summary function on a scpecific input files and export the results into a csv}
\usage{
csv_summary(input, output)
}
\arguments{
\item{input}{(character) path to input csv file}

\item{output}{(character) path to output summary file}
}
\description{
Apply summary function on a scpecific input files and export the results into a csv
}
\examples{

}
