% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issue_extract.R
\name{issue_extract}
\alias{issue_extract}
\title{Export GitHub issues as PDF Files}
\usage{
issue_extract(
  repo_url = NULL,
  start = 1,
  end = 10,
  export_folder = "exported_issues",
  cookies = NULL
)
}
\arguments{
\item{repo_url}{(character) URL of the GitHub repository as a character string.}

\item{start}{(numeric) First GitHub issue to be exported. Default is issue number 1.}

\item{end}{(numeric) Last GitHub issue to be exported. Default is issue number 10.}

\item{export_folder}{(character) Name of the folder that will be created to contain the output PDF files. Default is "exported_issues".}

\item{cookies}{(character) Optional file path to the cookies to load into the Chrome session.}
}
\value{
No return value, called for side effects
}
\description{
Exports specified GitHub issues as PDF files when given the URL of a GitHub repository. This function will export the first 10 issues as a default.
}
\examples{
\dontrun{
# Export GitHub issues #7080 through #7089 for the public `dplyr` repository 
issue_extract(repo_url = "https://github.com/tidyverse/dplyr",
              start = 7080,
              end = 7089,
              export_folder = "dplyr_issues")
}
\dontrun{
# Export GitHub issues #295 through #300 for a GitHub Enterprise repository
# First save your credentials as cookies
# See https://github.com/rstudio/chromote/blob/main/README.md#websites-that-require-authentication for more info

# Start new Chrome session
b <- ChromoteSession$new()
# Open interactive window
b$view()

# Navigate to a random Github Enterprise issue
# Make sure to log in to GitHub Enterprise
b$Page$navigate("https://github.nceas.ucsb.edu/LTER/lter-wg-scicomp/issues/278")

# Save credentials as cookies
cookies <- b$Network$getCookies()
saveRDS(cookies, "cookies.rds")

# Close the browser tab/window
b$close()

# After saving cookies, you should restart R
# Then read in the cookies and export the necessary issues 
issue_extract(repo_url = "https://github.nceas.ucsb.edu/LTER/lter-wg-scicomp",
              start = 295,
              end = 300,
              export_folder = "scicomp_issues",
              cookies = "cookies.rds")
}
}
