% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/molec_wt.R
\name{molec_wt}
\alias{molec_wt}
\title{Identify Molecular Weight for a Given Element}
\usage{
molec_wt(element = NULL)
}
\arguments{
\item{element}{(character/numeric) element name, symbol, or atomic number for which to retrieve molecular weight}
}
\value{
(numeric) molecular weight value for the relevant element
}
\description{
Identifies molecular weight for the specified element based on the element's name, its symbol, or its atomic number. Returns only the molecular weight as a numeric value.
}
\examples{
# Identify molecular weight for carbon by name
molec_wt(element = "Carbon")

# Identify molecular weight for hydrogen by atomic number
molec_wt(element = 1)

}
