% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlme_extract.R
\name{nlme_extract}
\alias{nlme_extract}
\title{Extract Estimates From `nlme::lme` Model Fit Object}
\usage{
nlme_extract(fit = NULL)
}
\arguments{
\item{fit}{(lme) Fit object returned by `nlme::lme`}
}
\description{
Extracts model terms, total sample size, estimates, standard errors, degrees of freedom, T values, and P values from a `nlme::lme` fit object
}
