% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel_extract.R
\name{output_bold}
\alias{output_bold}
\title{Extract bold formatting information to csv}
\usage{
output_bold(contents, formats, directory, sheet)
}
\arguments{
\item{contents}{(data frame) values stored in the Excel sheet}

\item{formats}{(data frame) values stored in the Excel sheet}

\item{directory}{(character) path to the folder where to save the csv files}

\item{sheet}{(character) name of the sheet to process}
}
\description{
Extract bold formatting information to csv
}
