% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_info_from_excel_workbook.R
\name{output_font}
\alias{output_font}
\title{Extract font type information to csv}
\usage{
output_font(contents, formats, directory, sheet)
}
\arguments{
\item{contents}{(data frame) values stored in the Excel sheet}

\item{formats}{(data frame) values stored in the Excel sheet}

\item{directory}{(character) path to the folder where to save the csv files}

\item{sheet}{(character) name of the sheet to process}
}
\description{
Extract font type information to csv
}
