% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_info_from_excel_workbook.R
\name{output_to_csv}
\alias{output_to_csv}
\title{Export csv files out of an Excel workbook including formatting and formulas}
\usage{
output_to_csv(filename)
}
\arguments{
\item{filename}{(character)  Name of the Excel workbook}
}
\description{
Export csv files out of an Excel workbook including formatting and formulas
}
