% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel_extract.R
\name{output_value}
\alias{output_value}
\title{Extract data information to csv}
\usage{
output_value(contents, formats, directory, sheet)
}
\arguments{
\item{contents}{(data frame) values stored in the Excel sheet}

\item{formats}{(data frame) values stored in the Excel sheet}

\item{directory}{(character) path to the folder where to save the csv files}

\item{sheet}{(character) name of the sheet to process}
}
\description{
Extract data information to csv
}
