% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv_import.R
\name{read.csv.gdrive}
\alias{read.csv.gdrive}
\title{Read file from Google Drive using the google id hash
addapted from https://github.com/sokole/ltermetacommunities/blob/master/examples/SOKOL-RScript-reading-from-google-drive.R}
\usage{
read.csv.gdrive(file_id_gdrive, skipper = 0,
  gdrive_url = "https://drive.google.com/uc?export=download&id=")
}
\arguments{
\item{file_id_gdrive}{(character) Google Drive csv file id A character}

\item{skipper}{(integer) Skip first lines An integer}

\item{gdrive_url}{(character) base URL An character}
}
\value{
data frame with the csv content A data frame
}
\description{
Read file from Google Drive using the google id hash
addapted from https://github.com/sokole/ltermetacommunities/blob/master/examples/SOKOL-RScript-reading-from-google-drive.R
}
\examples{
\dontrun{
my_data <- read.csv.gdrive("0B7AABlvKD6WjSTY3YUVKZ1AwLWs")
}

}
