% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv_import.R
\name{read_csv_gdrive}
\alias{read_csv_gdrive}
\title{Read file from Google Drive using the google id hash using readr::read_csv}
\usage{
read_csv_gdrive(file_id_gdrive, skipper = 0,
  gdrive_url = "https://drive.google.com/uc?export=download&id=")
}
\arguments{
\item{file_id_gdrive}{Google Drive csv file id A character}

\item{skipper}{Skip first lines An integer}

\item{gdrive_url}{base URL An character}
}
\value{
data frame with the csv content A data frame
}
\description{
Read file from Google Drive using the google id hash using readr::read_csv
}
\examples{
\dontrun{
my_data <- read_csv_gdrive("0B7AABlvKD6WjSTY3YUVKZ1AwLWs")
}

}
