% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_info_from_excel_workbook.R
\name{read_excel_allsheets}
\alias{read_excel_allsheets}
\title{Read all the sheets from an Excel workbook}
\usage{
read_excel_allsheets(filename)
}
\arguments{
\item{filename}{(character)  Name of the Excel workbook}
}
\description{
Read all the sheets from an Excel workbook
}
