% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel_extract.R
\name{read_full_excel}
\alias{read_full_excel}
\title{Read All Sheets from an Excel Workbook}
\usage{
read_full_excel(filename = NULL)
}
\arguments{
\item{filename}{(character) Name of the Excel workbook}
}
\description{
Retrieves all of the sheets in a given Microsfot Excel workbook and stores them as elements in a list. Note that the guts of this function were created by the developers of `readxl` and a tutorial-ized version of this function can be found here: https://readxl.tidyverse.org/articles/readxl-workflows.html (Note however that they didn't have it as a function)
}
