% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geospatial.R
\name{regrid_to_regular}
\alias{regrid_to_regular}
\title{Interpolate netCDF to WGS84}
\usage{
regrid_to_regular(
  netcdf_file,
  variable_to_extract,
  grid_resolution = 0.1,
  filling_value_flag = TRUE,
  lat = "TLONG",
  long = "TLONG"
)
}
\arguments{
\item{netcdf_file}{(character) path to netCDF file}

\item{variable_to_extract}{(character)}

\item{grid_resolution}{(numeric)}

\item{filling_value_flag}{(logical)}

\item{lat}{(atomic vector)}

\item{long}{(atomic vector)}
}
\value{
(rasterLayer) regridded raster
}
\description{
Accepts a netCDF file and re-projects it into a new rasterLayer that has been re-gridded into World Geodetic System 84 (WGS84). Note that this relies on the `raster` package which is becoming superseded by `stars`, `terra`, and `sf`
}
\examples{
\dontrun{
regrid_to_regular("cesm_co2sys_2005.nc", "OARG")
}

}
