% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_export.R
\name{stat_export}
\alias{stat_export}
\title{Export Summary Statistics from a Model Fit}
\usage{
stat_export(
  model_obj = NULL,
  model_type = "lmer",
  output_path = getwd(),
  output_name = paste0(model_type, "_", Sys.time(), "_model.csv"),
  est_dig = 2,
  se_dig = 2,
  df_dig = 2,
  t_dig = 2,
  p_dig = 4
)
}
\arguments{
\item{model_obj}{output of `lmerTest::lmer()`, `stats::lm()`, `stats::nls()`, `stats::t.test()`}

\item{model_type}{(character) One of "lmer", "lm", "nls" or "t-test"}

\item{output_path}{(character) File path to save file to}

\item{output_name}{(character) Desired name of file (defaults to model type and system time)}

\item{est_dig}{(numeric) Number of digits to round estimate to}

\item{se_dig}{(numeric) Number of digits to round standard error to}

\item{df_dig}{(numeric) Number of digits to round degrees of freedom to}

\item{t_dig}{(numeric) Number of digits to round T value to}

\item{p_dig}{(numeric) Number of digits to round P value to}
}
\description{
Accepts a model fit object for one of a set of models and saves a CSV of summary statistics to the desired location. Allows user to define the number of digits for each summary statistic if desired
}
