% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_extract.R
\name{stat_extract}
\alias{stat_extract}
\title{Extract Summary Statistics from Model Fit Object}
\usage{
stat_extract(mod_fit = NULL, traj_angle = "deg")
}
\arguments{
\item{mod_fit}{(lme, trajectory.analysis) Model fit object of supported class (see function description text)}

\item{traj_angle}{(character) Either "deg" or "rad" for whether trajectory analysis angle information should be extracted in degrees or radians. Only required if model is trajectory analysis}
}
\value{
(data.frame) Dataframe of core summary statistics for the given model
}
\description{
Accepts model fit object and extracts core statistical information. This includes P value, test statistic, degrees of freedom, etc. Currently accepts the following model types: `stats::t.test`, `stats::lm`, `stats_nls`, `nlme::lme`, `lmerTest::lmer`, `ecodist::MRM`, or `RRPP::trajectory.analysis`
}
