% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traj_extract.R
\name{traj_extract}
\alias{traj_extract}
\title{Extract Summary Statistics From `RRPP::trajectory.analysis` Model Fit Object}
\usage{
traj_extract(traj_mod = NULL, angle_type = "deg")
}
\arguments{
\item{traj_mod}{(trajectory.analysis) Object returned by `RRPP::trajectory.analysis`}

\item{angle_type}{(character) Either "deg" or "rad" for what units to use for trajectory analysis angle metric. See `?RRPP::summary.trajectory.analysis` for more information}
}
\value{
(data.frame) Flat table including various useful summary metrics for the three trajectory analysis metrics
}
\description{
Extracts summary tables for distance, shape, and angle results identified by `RRPP::trajectory.analysis`. Returns as a flat dataframe for convenience
}
