% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zoom_webinar_fix.R
\name{zoom_webinar_fix}
\alias{zoom_webinar_fix}
\title{Wrangle Zoom Webinar Report CSVs}
\usage{
zoom_webinar_fix(file, report_type = "attendance")
}
\arguments{
\item{file}{character, file name/path of csv output from Zoom webinar}

\item{report_type}{character, one of "attendance" or "survey"}
}
\value{
dataframe containing tidied (and machine-readable) webinar report. Also drops identifying information in attendance report (no such information in survey response).
}
\description{
Zoom webinars produce report CSVs of (1) attendance and (2) post-webinar survey responses. Both of these can be produced with human-readable header content before getting to participant data; unfortunately, these headers completely destroy R's ability to comprehend the column names for the relevant bits so a function was necessary to wrangle both types of report. Additionally, this function identifies the job sector of attendees based on their email domain
}
