% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarization.R
\name{categorical_frequency}
\alias{categorical_frequency}
\title{Compute categorical frequency}
\usage{
categorical_frequency(df, threshold_fract = 0.9)
}
\arguments{
\item{df}{A data frame containing the data to summarize}
}
\value{
summary statistics for categorical values
}
\description{
Compute categorical frequency
}
\examples{
\dontrun{
categorical_frequency(faimstracs_legacy)
}

}
