% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarization.R
\name{csv_summary}
\alias{csv_summary}
\title{Summarize and Export a CSV}
\usage{
csv_summary(input = NULL, output = NULL)
}
\arguments{
\item{input}{(character) path to input csv file}

\item{output}{(character) path to output summary file}
}
\description{
Applies `Hmisc::describe()` to a specific input CSV file and outputs the results into a new csv
}
