% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morpho_wiz.R
\name{morpho_wiz}
\alias{morpho_wiz}
\title{Pre-Morpho Wrangling}
\usage{
morpho_wiz(file = NULL, read_dir = getwd(), write_dir = getwd(), na_code = "")
}
\arguments{
\item{file}{character, name of file to check}

\item{read_dir}{character, path to file to check}

\item{write_dir}{character, path to write checked file to}

\item{na_code}{character, how missing values are coded in the data}
}
\description{
"Morpho" refers to dataONE's metadata software **NOT** the 2022 NCEAS working group request for proposals of the same name. See [here](https://old.dataone.org/software-tools/morpho) for the Morpho these functions are meant to deal with. Written by: Justin Kroes (jkroes14@ucsbalum.com | orcid.org/0000-0003-3257-2321)
}
