% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_info_from_excel_workbook.R
\name{output_sheet}
\alias{output_sheet}
\title{Wrapper function around the other functions to extract information from excel sheet into separate csv files}
\usage{
output_sheet(filename, directory, sheet, n)
}
\arguments{
\item{filename}{(character)  Name of the Excel workbook}

\item{directory}{(character)  Path to the output directory}

\item{sheet}{(character)  Excel sheet name}

\item{n}{(integer)  Sheet number}
}
\value{

}
\description{
Wrapper function around the other functions to extract information from excel sheet into separate csv files
}
