% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geospatial.R
\name{regrid_to_regular}
\alias{regrid_to_regular}
\title{Interpolate a netcdf file to a regular World Geodetic System (WGS84) grid}
\usage{
regrid_to_regular(
  netcdf_file,
  variable_to_extract,
  grid_resolution = 0.1,
  filling_value_flag = TRUE,
  lat = "TLONG",
  long = "TLONG"
)
}
\arguments{
\item{netcdf_file}{A character}

\item{variable_to_extract}{A character}

\item{grid_resolution}{A numeric}

\item{filling_value_flag}{A logical}

\item{lat}{An atomic vector}

\item{long}{An atomic vector}
}
\value{
(rasterLayer) regridded raster
}
\description{
Interpolate a netcdf file to a regular World Geodetic System (WGS84) grid
}
\examples{
\dontrun{regrid_to_regular("cesm_co2sys_2005.nc", "OARG")}

}
